+++ NEWS.md
@@ -5,6 +5,8 @@ Version 0.12.5 For Dev-C++ 7 Beta
  - enhancement: add option for setting astyle path
  - fix: wrong file wildcard (*.*) in linux
  - fix: open terminal in linux
+ - fix: wrong executable filename for source files in linux
+ - enhancement: console pauser for linux 
 
 Version 0.12.4 For Dev-C++ 7 Beta
  - change: add copyright infos to each source file